% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{server}
\alias{server}
\title{server for untidywindrose.}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{the required rshiny input object}

\item{output}{the required rshiny output object}

\item{session}{the optional rshiny session object}
}
\value{
NA
}
\description{
The Rshiny's server object for untidywindrose
}
