% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{ui}
\alias{ui}
\title{ui for untidywindrose.}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.
}
\usage{
ui
}
\value{
NA
}
\description{
The Rshiny's UI object for untidywindrose
}
\keyword{datasets}
