% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untidywindrose.R
\name{untidywindrose}
\alias{untidywindrose}
\title{untidywindrose}
\usage{
untidywindrose()
}
\value{
shiny application object
}
\description{
starts the untidywindrose shiny app
}
\details{
launches the untidywindrose shiny app
}
